Implementação de requisições realizadas com Token Bearer
Em Despesas e Categoria Services foi adicionado método GetByIdUsuario
Implementação de tratamento de Errro classe ErroInterceptor
Tentativa de correção design LayoutMaster

+++ src/shared/services/api/ReceitasService.ts
@@ -1,7 +1,8 @@
-import { Api } from "../axios-config";
+import  createApiInstance   from "../axios-config";
 import { Dayjs } from "dayjs";
 
-export interface IReceitaVO {
+const Api = createApiInstance();
+export interface IReceitaVM {
     id:number;
     idUsuario: number;
     idCategoria: Number;
@@ -10,9 +11,11 @@ export interface IReceitaVO {
     valor: number;
 }
 
-const getAll = async (): Promise<IReceitaVO[] | Error> => {
+const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+const getAll = async (): Promise<IReceitaVM[] | Error> => {
     try {
-        const { data } = await Api.get('/Receita');
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.get('/Receita', {headers: { Authorization: `Bearer ${accessToken}` }});
         if (data) {
             return data;
         }
@@ -24,11 +27,12 @@ const getAll = async (): Promise<IReceitaVO[] | Error> => {
     }
 };
 
-const getById = async (id: Number): Promise<IReceitaVO | Error> => {
-    try {
-        const { data } = await Api.get('/Receita/' + id);
+const getById = async (id: Number): Promise<IReceitaVM | Error> => {
+    try {        
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.get('/Receita/GetById/' + id, {headers: { Authorization: `Bearer ${accessToken}` }});
         if (data) {
-            return data.receita as IReceitaVO;
+            return data.receita as IReceitaVM;
         }
 
         return Error('Erro getById ao pesquisar receitas.');
@@ -38,9 +42,30 @@ const getById = async (id: Number): Promise<IReceitaVO | Error> => {
     }
 };
 
-const create = async (dados: Omit<IReceitaVO, 'id'>): Promise<any | Error> => {
+const getByIdUsuario = async (idUsuario: number): Promise<IReceitaVM[] | Error> => {
     try {
-        const { data } = await Api.post<IReceitaVO>('/Receita', dados );
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.get('/despesa/GetByIdUsuario/' + idUsuario, {
+            headers: {
+                 Authorization: `Bearer ${accessToken}` 
+                }
+            });
+        if (data) {
+            return data;
+        }
+
+        return Error('Erro getByIdUsuario ao listar Categorias.');
+    } catch (error) {
+        console.log(error);
+        return Error((error as { message: string }).message || 'Erro getByIdusuario ao listar Categorias.');
+    }
+};
+
+
+const create = async (dados: Omit<IReceitaVM, 'id'>): Promise<any | Error> => {
+    try {        
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.post<IReceitaVM>('/Receita', dados, {headers: { Authorization: `Bearer ${accessToken}` }} );
         if (data) {
             return data.id
         }
@@ -53,10 +78,11 @@ const create = async (dados: Omit<IReceitaVO, 'id'>): Promise<any | Error> => {
 };
 
 
-const updateById = async (id: number, dados: IReceitaVO): Promise<IReceitaVO | Error> => {
-    try {
-        dados.id = id;
-        const { data } = await Api.put<IReceitaVO>('/Receita', dados);
+const updateById = async (id: number, dados: IReceitaVM): Promise<IReceitaVM | Error> => {
+    try {        
+        dados.id = id;        
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.put<IReceitaVM>('/Receita', dados);
         if (data) {
             return data
         }
@@ -70,8 +96,9 @@ const updateById = async (id: number, dados: IReceitaVO): Promise<IReceitaVO | E
  };
 
 const deleteById = async (id: number): Promise<any | Error> => { 
-    try {
-        const { data } = await Api.delete('/Receita/' + id);
+    try {     
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');   
+        const { data } = await Api.delete('/Receita/' + id, {headers: { Authorization: `Bearer ${accessToken}` }});
         if (data.message) {
             return Boolean(data)
         }
@@ -87,6 +114,7 @@ const deleteById = async (id: number): Promise<any | Error> => {
 export const ReceitasService = {
     getAll,
     getById,
+    getByIdUsuario,
     create,
     updateById,
     deleteById
