+++ src/shared/services/api/UsuariosService.ts
@@ -1,15 +1,17 @@
-import { Api } from "../axios-config";
+import  createApiInstance   from "../axios-config";
+const Api = createApiInstance();
 
-export interface IUsuarioVO {
+export interface IUsuarioVM {
     Id:number;
     Nome: string;
     Telefone: string;
     Email: string;
 }
 
-const getAll = async (): Promise<IUsuarioVO[] | Error> => {
+const getAll = async (): Promise<IUsuarioVM[] | Error> => {
     try {
-        const { data } = await Api.get('/Usuario');
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.get('/Usuario', {headers: { Authorization: `Bearer ${accessToken}` }});        
         if (data) {
             return data;
         }
@@ -21,9 +23,10 @@ const getAll = async (): Promise<IUsuarioVO[] | Error> => {
     }
 };
 
-const getById = async (id: Number): Promise<IUsuarioVO | Error> => {
+const getById = async (id: Number): Promise<IUsuarioVM | Error> => {
     try {
-        const { data } = await Api.get('/Usuario/$(id)');
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.get('/Usuario/$(id)', {headers: { Authorization: `Bearer ${accessToken}` }});
         if (data) {
             return data;
         }
@@ -35,9 +38,10 @@ const getById = async (id: Number): Promise<IUsuarioVO | Error> => {
     }
 };
 
-const create = async (dados: Omit<IUsuarioVO, 'id'>): Promise<number | Error> => {
+const create = async (dados: Omit<IUsuarioVM, 'id'>): Promise<number | Error> => {
     try {
-        const { data } = await Api.post<IUsuarioVO>('/Usuario', dados );
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.post<IUsuarioVM>('/Usuario', dados, {headers: { Authorization: `Bearer ${accessToken}` }} );
         if (data) {
             return data.Id
         }
@@ -50,10 +54,11 @@ const create = async (dados: Omit<IUsuarioVO, 'id'>): Promise<number | Error> =>
 };
 
 
-const updateById = async (id: number, dados: IUsuarioVO): Promise<IUsuarioVO | Error> => {
+const updateById = async (id: number, dados: IUsuarioVM): Promise<IUsuarioVM | Error> => {
     try {
         dados.Id = id;
-        const { data } = await Api.put<IUsuarioVO>('/Usuario', dados);
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.put<IUsuarioVM>('/Usuario', dados, {headers: { Authorization: `Bearer ${accessToken}` }});
         if (data) {
             return data
         }
@@ -68,7 +73,8 @@ const updateById = async (id: number, dados: IUsuarioVO): Promise<IUsuarioVO | E
 
 const deleteById = async (id: number): Promise<void | Error> => { 
     try {
-        const { data } = await Api.delete('/Usuario/$(id)');
+        const accessToken = localStorage.getItem('@dpApiAccess')?.replaceAll('"', '');
+        const { data } = await Api.delete('/Usuario/$(id)', {headers: { Authorization: `Bearer ${accessToken}` }});
         if (data) {
             return data.id
         }
@@ -78,7 +84,6 @@ const deleteById = async (id: number): Promise<void | Error> => {
         console.log(error);
         return Error((error as { message: string }).message || 'Erro ao deletar registro de Usuario.');
     }
-
 };
 
 export const UsuariosService = {
